unify colors
